Improve partner messaging UX
